feat: add file and video support
Now, `GoCQHttp` provides two different type of the file upload:

+ user: should call `/upload_private_file`
+ group: should call `/upload_group_file`

It's not difficult to combine this into one function `coolq_send_file`.
However, the tricky part is two things.

Because, `efb-qq-salve` does not support dynamic supported messages, so
this commit is highly mixed with the
[PR](https://github.com/ehForwarderBot/efb-qq-slave/pull/119).

And there is also a problem that the file would cause `NetworkError`
exception. `aiocqhttp` does exactly provides the `api_timeout_sec`,
however `aiocqhttp` does not use this variable in the `HttpApi`.
See [PR](https://github.com/nonebot/aiocqhttp/pull/71).

It's tricky to determine the timeout, because there are so many aspects:
the network and the file size. So this commit allows user to specify
the coolq api timeout.

However, for the video, the name could not be consistent, this is the
restrictions from the master, see [issue](https://github.com/ehForwarderBot/efb-telegram-master/issues/136)

close #61 close #64